requirements.txt
selenium==3.141.0
msedge_selenium_tools==3.141.3
